chore: format

+++ contracts/core/EntryPointSimulations.sol
@@ -14,8 +14,7 @@ import "../interfaces/IEntryPointSimulations.sol";
  */
 contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
     // solhint-disable-next-line var-name-mixedcase
-    AggregatorStakeInfo private NOT_AGGREGATED =
-        AggregatorStakeInfo(address(0), StakeInfo(0, 0));
+    AggregatorStakeInfo private NOT_AGGREGATED = AggregatorStakeInfo(address(0), StakeInfo(0, 0));
 
     /**
      * simulation contract should not be deployed, and specifically, accounts should not trust
@@ -26,33 +25,21 @@ contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
     }
 
     /// @inheritdoc IEntryPointSimulations
-    function simulateValidation(
-        UserOperation calldata userOp
-    ) external returns (ValidationResult memory) {
+    function simulateValidation(UserOperation calldata userOp) external returns (ValidationResult memory) {
         UserOpInfo memory outOpInfo;
 
         _simulationOnlyValidations(userOp);
-        (
-            uint256 validationData,
-            uint256 paymasterValidationData
-        ) = _validatePrepayment(0, userOp, outOpInfo);
-        StakeInfo memory paymasterInfo = _getStakeInfo(
-            outOpInfo.mUserOp.paymaster
-        );
+        (uint256 validationData, uint256 paymasterValidationData) = _validatePrepayment(0, userOp, outOpInfo);
+        StakeInfo memory paymasterInfo = _getStakeInfo(outOpInfo.mUserOp.paymaster);
         StakeInfo memory senderInfo = _getStakeInfo(outOpInfo.mUserOp.sender);
         StakeInfo memory factoryInfo;
         {
             bytes calldata initCode = userOp.initCode;
-            address factory = initCode.length >= 20
-                ? address(bytes20(initCode[0:20]))
-                : address(0);
+            address factory = initCode.length >= 20 ? address(bytes20(initCode[0:20])) : address(0);
             factoryInfo = _getStakeInfo(factory);
         }
 
-        ValidationData memory data = Helpers._intersectTimeRange(
-            validationData,
-            paymasterValidationData
-        );
+        ValidationData memory data = Helpers._intersectTimeRange(validationData, paymasterValidationData);
         address aggregator = data.aggregator;
         bool sigFailed = aggregator == address(1);
         ReturnInfo memory returnInfo = ReturnInfo(
@@ -66,37 +53,21 @@ contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
 
         AggregatorStakeInfo memory aggregatorInfo = NOT_AGGREGATED;
         if (aggregator != address(0) && aggregator != address(1)) {
-            aggregatorInfo = AggregatorStakeInfo(
-                aggregator,
-                _getStakeInfo(aggregator)
-            );
+            aggregatorInfo = AggregatorStakeInfo(aggregator, _getStakeInfo(aggregator));
         }
-        return
-            ValidationResult(
-                returnInfo,
-                senderInfo,
-                factoryInfo,
-                paymasterInfo,
-                aggregatorInfo
-            );
+        return ValidationResult(returnInfo, senderInfo, factoryInfo, paymasterInfo, aggregatorInfo);
     }
 
     /// @inheritdoc IEntryPointSimulations
-    function simulateHandleOp(
-        UserOperation calldata op,
-        address target,
-        bytes calldata targetCallData
-    ) external nonReentrant returns (ExecutionResult memory) {
+    function simulateHandleOp(UserOperation calldata op, address target, bytes calldata targetCallData)
+        external
+        nonReentrant
+        returns (ExecutionResult memory)
+    {
         UserOpInfo memory opInfo;
         _simulationOnlyValidations(op);
-        (
-            uint256 validationData,
-            uint256 paymasterValidationData
-        ) = _validatePrepayment(0, op, opInfo);
-        ValidationData memory data = Helpers._intersectTimeRange(
-            validationData,
-            paymasterValidationData
-        );
+        (uint256 validationData, uint256 paymasterValidationData) = _validatePrepayment(0, op, opInfo);
+        ValidationData memory data = Helpers._intersectTimeRange(validationData, paymasterValidationData);
 
         numberMarker();
         uint256 paid = _executeUserOp(0, op, opInfo);
@@ -106,29 +77,12 @@ contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
         if (target != address(0)) {
             (targetSuccess, targetResult) = target.call(targetCallData);
         }
-        return
-            ExecutionResult(
-                opInfo.preOpGas,
-                paid,
-                data.validAfter,
-                data.validUntil,
-                targetSuccess,
-                targetResult
-            );
+        return ExecutionResult(opInfo.preOpGas, paid, data.validAfter, data.validUntil, targetSuccess, targetResult);
     }
 
-    function _simulationOnlyValidations(
-        UserOperation calldata userOp
-    ) internal view {
-        try
-            this._validateSenderAndPaymaster(
-                userOp.initCode,
-                userOp.sender,
-                userOp.paymasterAndData
-            )
-        // solhint-disable-next-line no-empty-blocks
-        {
-
+    function _simulationOnlyValidations(UserOperation calldata userOp) internal view {
+        try this._validateSenderAndPaymaster(userOp.initCode, userOp.sender, userOp.paymasterAndData) {
+            // solhint-disable-next-line no-empty-blocks
         } catch Error(string memory revertReason) {
             if (bytes(revertReason).length != 0) {
                 revert FailedOp(0, revertReason);
@@ -143,11 +97,10 @@ contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
      * @param sender           - The sender address.
      * @param paymasterAndData - The paymaster address followed by the token address to use.
      */
-    function _validateSenderAndPaymaster(
-        bytes calldata initCode,
-        address sender,
-        bytes calldata paymasterAndData
-    ) external view {
+    function _validateSenderAndPaymaster(bytes calldata initCode, address sender, bytes calldata paymasterAndData)
+        external
+        view
+    {
         if (initCode.length == 0 && sender.code.length == 0) {
             // it would revert anyway. but give a meaningful message
             revert("AA20 account not deployed");
@@ -165,10 +118,8 @@ contract EntryPointSimulations is EntryPoint, IEntryPointSimulations {
 
     //make sure depositTo cost is more than normal EntryPoint's cost.
     // empiric test showed that without this wrapper, simulation depositTo costs less..
-    function depositTo(
-        address account
-    ) public payable override(IStakeManager, StakeManager) {
-        uint x;
+    function depositTo(address account) public payable override(IStakeManager, StakeManager) {
+        uint256 x;
         assembly {
             //some silly code to waste ~200 gas
             x := exp(mload(0), 100)
