+++ contracts/test/TestSignatureAggregator.sol
@@ -12,31 +12,33 @@ import "../samples/SimpleAccount.sol";
  * the aggregated signature is the SUM of the nonce fields..
  */
 contract TestSignatureAggregator is IAggregator {
-
     /// @inheritdoc IAggregator
     function validateSignatures(UserOperation[] calldata userOps, bytes calldata signature) external pure override {
-        uint sum = 0;
-        for (uint i = 0; i < userOps.length; i++) {
-            uint nonce = userOps[i].nonce;
+        uint256 sum = 0;
+        for (uint256 i = 0; i < userOps.length; i++) {
+            uint256 nonce = userOps[i].nonce;
             sum += nonce;
         }
         require(signature.length == 32, "TestSignatureValidator: sig must be uint");
-        (uint sig) = abi.decode(signature, (uint));
+        (uint256 sig) = abi.decode(signature, (uint256));
         require(sig == sum, "TestSignatureValidator: aggregated signature mismatch (nonce sum)");
     }
 
     /// @inheritdoc IAggregator
-    function validateUserOpSignature(UserOperation calldata)
-    external pure returns (bytes memory) {
+    function validateUserOpSignature(UserOperation calldata) external pure returns (bytes memory) {
         return "";
     }
 
     /**
      * dummy test aggregator: sum all nonce values of UserOps.
      */
-    function aggregateSignatures(UserOperation[] calldata userOps) external pure returns (bytes memory aggregatedSignature) {
-        uint sum = 0;
-        for (uint i = 0; i < userOps.length; i++) {
+    function aggregateSignatures(UserOperation[] calldata userOps)
+        external
+        pure
+        returns (bytes memory aggregatedSignature)
+    {
+        uint256 sum = 0;
+        for (uint256 i = 0; i < userOps.length; i++) {
             sum += userOps[i].nonce;
         }
         return abi.encode(sum);
